requirements.txt
alembic==1.3.2
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Mako==1.1.0
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.18.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
twilio==6.35.1
urllib3==1.25.7
Werkzeug==0.16.0
pyserial==3.4